==> notebooks/jason/0_data_preprocessing.ipynb <==
# Preprocessing

All four types of inputs

1. flow and transaction network data.
2. Shapefile of SA2.
3. Shapefile of road networks.
4. economic indicators, e.g. jobs, employment, etc.

Three inputs need to be processed.

1. flow and transaction network data.
2. Shapefile of SA2.
3. Shapefile of road networks. (to be done)

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import geopandas as gpd
# import geoplot

## 1. Flow and transaction network data

# mount_path = "/mnt/c/Users/jason/Dropbox (MIT)/"
mount_path = "/Users/shenhaowang/Dropbox (MIT)/project_media_lab_South_Australia/"

# read data

### flow
flow_df = pd.read_csv(mount_path + "/SA data/dataSA/flows_sa2_months_2018-02-01.csv")
print("flow shape:", flow_df.shape)
print(flow_df.head(10))
print()

### transaction 1
trans_age_df = pd.read_csv(mount_path + "/SA data/dataSA/transaction_age_bins.csv")
print("trans_age_df: ", trans_age_df.shape)
print(trans_age_df.head(10))
print()

### transaction 2
trans_mcc_df = pd.read_csv(mount_path + "/SA data/dataSA/transaction_mcc.csv")
print("trans_mcc_df: ", trans_mcc_df.shape)
print(trans_mcc_df.head(10))
print()

# Change sa2 values to string
flow_df['sa2'] = flow_df.sa2.astype(str)
flow_df['agent_home_sa2'] = [x[:-2] for x in flow_df['agent_home_sa2'].astype(str)]

trans_age_df['source_sa2'] = trans_age_df['source_sa2'].astype(str)
trans_age_df['target_sa2'] = trans_age_df['target_sa2'].astype(str)

trans_mcc_df['source_sa2'] = trans_mcc_df['source_sa2'].astype(str)
trans_mcc_df['target_sa2'] = trans_mcc_df['target_sa2'].astype(str)

# Shapes
print("flow shape:", flow_df.shape)
print("trans_age_df: ", trans_age_df.shape)
print("trans_mcc_df: ", trans_mcc_df.shape)

# describe
print(flow_df.dtypes)

print(trans_age_df.dtypes)

print(trans_mcc_df.dtypes)

#### Q: what are the 'Cell Size Limit', 'OUTST', and 'nan'? Can I just drop the values?
A: Cell Size Limit - Less than 40 people, (very small number). Nan - zero. 

# check a few columns' info
# Note: total number of SA2 in South Australia is 172.
# Many residents' locations from the credit card data are outside South Australia region.
print("--- number of sa2 in flow_df ---")
print("Flow SA2: ", len(np.unique(flow_df.sa2, return_counts = True)[0])) # 
# print(np.unique(flow_df.sa2, return_counts = True)) # 
print("Agent home SA2: ", len(np.unique(flow_df.agent_home_sa2, return_counts = True)[0])) # 
# print(np.unique(flow_df.agent_home_sa2, return_counts = True)) # 

print("--- sa2 in trans_age_df ---")
print("Source sa2: ", len(np.unique(trans_age_df.source_sa2, return_counts = True)[0])) # 
# print(np.unique(trans_age_df.source_sa2, return_counts = True)) # 
print("Target sa2: ", len(np.unique(trans_age_df.target_sa2, return_counts = True)[0])) # 
# print(np.unique(trans_age_df.target_sa2, return_counts = True)) # 

print("--- sa2 in trans_mcc_df ---")
print("Source sa2: ", len(np.unique(trans_mcc_df.source_sa2, return_counts = True)[0])) # 
# print(np.unique(trans_mcc_df.source_sa2, return_counts = True)) # 
print("Target sa2: ", len(np.unique(trans_mcc_df.target_sa2, return_counts = True)[0])) # 
# print(np.unique(trans_mcc_df.target_sa2, return_counts = True)) # 

print("--- activity types in trans_mcc_df ---")
# print(np.unique(trans_mcc_df.mcc, return_counts = True)) # it is activity types.



# process the three dfs to keep only the agents in SA areas

# flow
south_australia_sa4_set = ['401','402','403','404','405','406','407']
flow_df = flow_df.loc[np.array([x[:3] in south_australia_sa4_set for x in flow_df.agent_home_sa2])]
flow_df = flow_df.loc[np.array([x[:3] in south_australia_sa4_set for x in flow_df.sa2])]

flow_df


# flow for adelaide
adelaide_sa4_set = ['401','402','403','404']

flow_adelaide_df = flow_df.loc[np.array([x[:3] in adelaide_sa4_set for x in flow_df.agent_home_sa2])]
flow_adelaide_df = flow_adelaide_df.loc[np.array([x[:3] in adelaide_sa4_set for x in flow_adelaide_df.sa2])]

flow_adelaide_df

# transaction age
invalid_value_list = ['Cell Size Limit', 'nan', 'OUTST']
trans_age_df=trans_age_df.loc[~trans_age_df.source_sa2.isin(invalid_value_list)]
trans_age_df=trans_age_df.loc[~trans_age_df.target_sa2.isin(invalid_value_list)]

trans_age_df


# transaction age for adelaide
trans_age_adelaide_df = trans_age_df.loc[np.array([x[:3] in adelaide_sa4_set for x in trans_age_df.source_sa2])]
trans_age_adelaide_df = trans_age_adelaide_df.loc[np.array([x[:3] in adelaide_sa4_set for x in trans_age_adelaide_df.target_sa2])]

trans_age_adelaide_df


# transaction mcc
# remove the invalid values in source_sa2 and target_sa2
invalid_value_list = ['Cell Size Limit', 'nan', 'OUTST']
trans_mcc_df=trans_mcc_df.loc[~trans_mcc_df.source_sa2.isin(invalid_value_list)]
trans_mcc_df=trans_mcc_df.loc[~trans_mcc_df.target_sa2.isin(invalid_value_list)]

trans_mcc_df


# transaction mcc for adelaide
trans_mcc_adelaide_df = trans_mcc_df.loc[np.array([x[:3] in adelaide_sa4_set for x in trans_mcc_df.source_sa2])]
trans_mcc_adelaide_df = trans_mcc_adelaide_df.loc[np.array([x[:3] in adelaide_sa4_set for x in trans_mcc_adelaide_df.target_sa2])]

trans_mcc_adelaide_df


#### save cleaned files

trans_mcc_df.to_pickle("../../data_process/trans_mcc_df.pkl")

trans_age_df.to_pickle("../../data_process/trans_age_df.pkl")

flow_df.to_pickle("../../data_process/flow_df.pkl")

flow_adelaide_df.to_pickle("../../data_process/flow_adelaide_df.pkl")

trans_age_adelaide_df.to_pickle("../../data_process/trans_age_adelaide_df.pkl")

trans_mcc_adelaide_df.to_pickle("../../data_process/trans_mcc_adelaide_df.pkl")

## 2. Spatial shapefiles

# Read SA2
sa2_shape = gpd.read_file(mount_path + "/SA data/dataSA/sa2/SA2_2016_AUST.shp")


# Basic info of sa2 for the whole Australia
# sa2 shapefile is quite self-explainary. It includes the names and ids of sa2, sa3, and sa4 areas.
print(type(sa2_shape))
print(sa2_shape.head(10))
print(sa2_shape.dtypes)


#### Q: Which part is the focus area? Should we model full South Australia or only Adelaide? A: If not very expensive, we can model the full South Australia.
#### Q: Is this processing correct? The SA4 indexes of South Australia and Adelaide. A: Right.

# Keep SA areas.
# Info: file:///Users/shenhaowang/Downloads/StatePublicHealthPlan_Final.pdf (page 32)
south_australia_sa4_set = ['401','402','403','404','405','406','407']
sa2_south_au = sa2_shape.loc[sa2_shape.SA4_CODE16.isin(south_australia_sa4_set)]
print(sa2_south_au.shape) # 172 SA2 regions are left.


# Keep adelaide areas.
adelaide_sa4_set = ['401','402','403','404']
sa2_adelaide = sa2_shape.loc[sa2_shape.SA4_CODE16.isin(adelaide_sa4_set)]
print(sa2_adelaide.shape)


sa2_south_au.to_file('../../data_process/shapefiles/sa2_south_au.shp')

sa2_adelaide.to_file('../../data_process/shapefiles/sa2_adelaide.shp')

# 3. Road networks

# Source: https://data.sa.gov.au/data/dataset/roads/resource/dbd6cc0f-317e-4fc7-b734-b81a518a3f00
sa2_roads = gpd.read_file(mount_path + "/SA data/dataSA/roads/Roads_GDA2020.shp")

print("shape of the road network is: ", sa2_roads.shape)

pd.set_option('display.max_columns', None)
sa2_roads.head(5)

# check metadata in the data folder: Roads_GDA2020_metadata.html
for col in sa2_roads.columns:
    print(col)

sa2_roads.loc[:15,'class']

# focus on only the class variable. 
# Drop 8 null rows.
sa2_roads = sa2_roads.loc[~sa2_roads['class'].isna(),]


np.unique(sa2_roads['class'], return_counts=True)

# separate three types of roads (class = LOCL, HWY, and UND)
sa2_roads_LOCL = sa2_roads.loc[sa2_roads['class'] == 'LOCL', :]
sa2_roads_HWY = sa2_roads.loc[sa2_roads['class'] == 'HWY', :]
sa2_roads_UND = sa2_roads.loc[sa2_roads['class'] == 'UND', :]

# save
# take a while to save...
sa2_roads.to_file("../../data_process/shapefiles/sa2_roads.shp")

# save three categories of roads
sa2_roads_LOCL.to_file("../../data_process/shapefiles/sa2_roads_LOCL.shp")
sa2_roads_HWY.to_file("../../data_process/shapefiles/sa2_roads_HWY.shp")
sa2_roads_UND.to_file("../../data_process/shapefiles/sa2_roads_UND.shp")

## 4. Jobs and income

jobs_all = pd.read_csv(mount_path + "/SA data/dataSA/SA2_Jobs_All_Jobs_and_Income.csv")
jobs_industries = pd.read_csv(mount_path + "/SA data/dataSA/SA2_Jobs_In_Australia_Employee_Jobs_and_Income.csv")

# process the dataframes' names. Removing the weird space before each column name.

new_idx = []

for col in jobs_industries.columns:
    if col[0]==' ':
        new_idx.append(col[1:])
    else:
        new_idx.append(col)

jobs_industries.columns = new_idx
print(jobs_industries.columns)

new_idx = []

for col in jobs_all.columns:
    if col[0]==' ':
        new_idx.append(col[1:])
    else:
        new_idx.append(col)

jobs_all.columns = new_idx
print(jobs_all.columns)


# data type of sa2 index
jobs_all['sa2_code16'] = jobs_all['sa2_code16'].astype('str')
jobs_industries['sa2_code16'] = jobs_industries['sa2_code16'].astype('str')

#### save files

jobs_all.to_pickle("../../data_process/jobs_all.pkl")

jobs_industries.to_pickle("../../data_process/jobs_industries.pkl")
