==> notebooks/jason/3-degree_df_merge.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import geopandas as gpd
import geoplot
import pickle
import geoplot.crs as gcrs
import statsmodels.api as sm

node_df = pd.read_pickle('../../data_process/node_specific_df.pickle')
degree_df = pd.read_pickle('../../data_process/degree_df.pickle')
edge_df = pd.read_pickle('../../data_process/edge_specific_df.pickle')

node_df.head()

node_df.columns

degree_df.head()

print(node_df.columns)
print(degree_df.columns)

node_degree_df = node_df.merge(degree_df, how="left", on="SA2_MAIN16")

node_degree_df.head()

node_degree_df.to_pickle('../../data_process/node_degree_df.pickle')

edge_df.head()

edge_df["origin"].values

origin_dest = list(zip(edge_df["origin"].values, edge_df["destination"].values))

len(origin_dest)

len(set(origin_dest))

edge_degree_df = pd.DataFrame(columns=["sa2_code16_x", "sa2_code16_y", "num_nodes_x", 
                                  "num_1degree_x", "num_2degree_x", "num_3degree_x", "num_4degree_x",
                                    "num_greater5degree_x",
                                      "num_nodes_y", 
                                  "num_1degree_y", "num_2degree_y", "num_3degree_y", "num_4degree_y",
                                    "num_greater5degree_y"])

edge_degree_df.head()

import pickle

with open("../../data_process/OD_full_path_dic.pickle", 'rb') as f:
    OD_full_path = pickle.load(f)

degree_df.columns

degree_df

sa_to_i = {}
i_to_sa = {}
sa_to_data = {}
for i, row in degree_df.iterrows():
    print(i)
    i_to_sa[i] = row["SA2_MAIN16"]
    sa_to_i[row["SA2_MAIN16"]] = i
    sa_to_data[row["SA2_MAIN16"]] = row[['num_nodes', 'num_1degree','num_2degree', 'num_3degree', 'num_4degree', 'num_greater5degree']]

for o,d in origin_dest:
    if o != d:
        o_data = degree_df[degree_df["SA2_MAIN16"]==o]
        d_data = degree_df[degree_df["SA2_MAIN16"]==d]
        
        num_nodes_pth = 0
        num_1degree_pth = 0
        num_2degree_pth = 0
        num_3degree_pth = 0
        num_4degree_pth = 0
        num_greater5degree_pth = 0
        oid = sa_to_i[o]
        did = sa_to_i[d]
        for i in OD_full_path[(oid,did)]:
            sa = i_to_sa[i]
            num_nodes_pth += float(sa_to_data[sa][0])
            num_1degree_pth += float(sa_to_data[sa][1])
            num_2degree_pth += float(sa_to_data[sa][2])
            num_3degree_pth += float(sa_to_data[sa][3])
            num_4degree_pth += float(sa_to_data[sa][4])
            num_greater5degree_pth += float(sa_to_data[sa][5])
                
        
        
        num_nodes_x = float(o_data["num_nodes"].iloc[0])
        num_1degree_x = float(o_data["num_1degree"].iloc[0])
        num_2degree_x = float(o_data["num_2degree"].iloc[0])
        num_3degree_x = float(o_data["num_3degree"].iloc[0])
        num_4degree_x = float(o_data["num_4degree"].iloc[0])
        num_greater5degree_x = float(o_data["num_greater5degree"].iloc[0])
        
        num_nodes_y = float(d_data["num_nodes"].iloc[0])
        num_1degree_y = float(d_data["num_1degree"].iloc[0])
        num_2degree_y = float(d_data["num_2degree"].iloc[0])
        num_3degree_y = float(d_data["num_3degree"].iloc[0])
        num_4degree_y = float(d_data["num_4degree"].iloc[0])
        num_greater5degree_y = float(d_data["num_greater5degree"].iloc[0])
        
        
    else:
        o_data = degree_df[degree_df["SA2_MAIN16"]==o]
        d_data = degree_df[degree_df["SA2_MAIN16"]==d]
        num_nodes_x = num_nodes_y = num_nodes_pth = float(o_data["num_nodes"].iloc[0])
        num_1degree_x = num_1degree_y = num_1degree_pth = float(o_data["num_1degree"].iloc[0])
        num_2degree_x = num_2degree_y = num_2degree_pth = float(o_data["num_2degree"].iloc[0])
        num_3degree_x = num_3degree_y = num_3degree_pth = float(o_data["num_3degree"].iloc[0])
        num_4degree_x = num_4degree_y = num_4degree_pth = float(o_data["num_4degree"].iloc[0])
        num_greater5degree_x = num_greater5degree_y = num_greater5degree_pth = float(o_data["num_greater5degree"].iloc[0])
    
    edge_degree_df = edge_degree_df.append({"sa2_code16_x": o, "sa2_code16_y":d ,"num_nodes_x":num_nodes_x, 
                                  "num_1degree_x":num_1degree_x, "num_2degree_x":num_2degree_x, 
                                  "num_3degree_x":num_3degree_x, "num_4degree_x":num_4degree_x,
                                  "num_greater5degree_x":num_greater5degree_x,
                                  "num_nodes_y":num_nodes_y, 
                                  "num_1degree_y":num_1degree_y, "num_2degree_y":num_2degree_y, 
                                  "num_3degree_y":num_3degree_y, "num_4degree_y":num_4degree_y,
                                  "num_greater5degree_y":num_greater5degree_y,
                                  "num_nodes_pth":num_nodes_pth,
                                  "num_1degree_pth":num_1degree_pth,
                                  "num_2degree_pth":num_2degree_pth,
                                  "num_3degree_pth":num_3degree_pth,
                                  "num_4degree_pth":num_4degree_pth,
                                   "num_greater5degree_pth":num_greater5degree_pth        },
                                ignore_index=True)
    

edge_degree_df.head()

edge_degree_df = edge_df.merge(edge_degree_df, how="left", on=["sa2_code16_x","sa2_code16_y"])

edge_degree_df.head()

edge_degree_df.to_pickle('../../data_process/edge_degree_df.pickle')
