==> notebooks/jason/3_data_merge.ipynb <==
# Data Prepare

Outputs: two dataframes - edge and node specific data sets.





import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import geopandas as gpd
import geoplot
import pickle
import geoplot.crs as gcrs

import statsmodels.api as sm

mount_path = "/mnt/c/Users/jason/Dropbox (MIT)/"

# read dfs
trans_mcc_adelaide_df = pd.read_pickle("../../data_process/trans_mcc_adelaide_df.pkl")
trans_age_adelaide_df = pd.read_pickle("../../data_process/trans_age_adelaide_df.pkl")
flow_adelaide_df = pd.read_pickle("../../data_process/flow_adelaide_df.pkl")

# read spatial files
sa2_south_au = gpd.read_file("../../data_process/shapefiles/sa2_south_au.shp")
sa2_adelaide = gpd.read_file('../../data_process/shapefiles/sa2_adelaide.shp')

# read road networks
sa2_roads = gpd.read_file("../../data_process/shapefiles/sa2_roads.shp")

# read job and income data
jobs_all=pd.read_pickle("../../data_process/jobs_all.pkl")
jobs_industries=pd.read_pickle("../../data_process/jobs_industries.pkl")

# read google api info
with open(mount_path + '/SA data/dataSA/OD_Google_API_raw.pickle', 'rb') as w:
    OD_google_raw = pickle.load(w)

with open(mount_path + '/SA data/dataSA/OD_Google_API_With_Map_Info.pickle', 'rb') as w:
    OD_google_with_map = pickle.load(w)


# read OD dictionary
with open("../../data_process/OD_full_path_dic.pickle", 'rb') as f:
    OD_full_path=pickle.load(f)

# read road class shp
sa2_adelaide_road_proj = gpd.read_file("../../data_process/shapefiles/sa2_adelaide_proj_road_class.shp")


# 
sa2_adelaide_road_proj

jobs_all.columns

jobs_all.shape

jobs_all.head(10)

# 
jobs_all_sub = jobs_all[['num_jobs_000_persons', 'sa2_code16', 'median_income_per_job_aud_persons']]
jobs_all_sub

# 
flow_adelaide_df.columns

# 
sa2_adelaide_road_proj.head()

# Prepare Network Data

# replace col names
flow_adelaide_df.rename(columns={'agent_home_sa2':'origin','sa2':'destination'}, inplace=True)
flow_adelaide_df


# init
flow_adelaide_df['OD'] = ''
flow_adelaide_df['OD'] = flow_adelaide_df['origin'] + flow_adelaide_df['destination']
flow_adelaide_df.groupby(by='OD').sum() # no repetition. 


# reindex
flow_adelaide_df.index = np.arange(flow_adelaide_df.shape[0])
flow_adelaide_df

# create ten columns here.
road_attribute_names_list = ['class_ART', 'class_BUS', 'class_COLL',
                             'class_FREE', 'class_HWY', 'class_LOCL', 'class_SUBA', 'class_TRK2',
                             'class_TRK4', 'class_UND']
flow_adelaide_df[road_attribute_names_list] = 0.0


# add the road attributes on the shortest path to the flow_adelaide_df.
# time cost: 3-5 mins?
for idx in np.arange(flow_adelaide_df.shape[0]):
    origin = flow_adelaide_df.loc[idx, 'origin']
    destination = flow_adelaide_df.loc[idx, 'destination']
    o_idx = sa2_adelaide.index[sa2_adelaide.SA2_MAIN16==origin].tolist()[0]
    d_idx = sa2_adelaide.index[sa2_adelaide.SA2_MAIN16==destination].tolist()[0]
#     print(o_idx,d_idx)
    
    try:
        # OD_full_path might not have all the shortest path...
        idx_list_on_shortest_path = OD_full_path[(o_idx, d_idx)]
        for node_on_shortest_path in idx_list_on_shortest_path:
            flow_adelaide_df.loc[idx, road_attribute_names_list] += sa2_adelaide_road_proj.loc[node_on_shortest_path, road_attribute_names_list]        
    except KeyError as error:
        pass


flow_adelaide_df.shape

flow_adelaide_df.head(20)

# add the job information to flow dataframe.
# origin
flow_adelaide_df=flow_adelaide_df.merge(jobs_all_sub, left_on='origin', right_on='sa2_code16', how = 'left')
flow_adelaide_df=flow_adelaide_df.rename(columns={'num_jobs_000_persons':'num_jobs_000_persons_origin', 'median_income_per_job_aud_persons':'median_income_per_job_aud_origin'})

# destination
flow_adelaide_df=flow_adelaide_df.merge(jobs_all_sub, left_on='destination', right_on='sa2_code16', how = 'left')
flow_adelaide_df=flow_adelaide_df.rename(columns={'num_jobs_000_persons':'num_jobs_000_persons_destination', 'median_income_per_job_aud_persons':'median_income_per_job_aud_destination'})

flow_adelaide_df

# augment the travel time and distance information to flow_adelaide_df
flow_adelaide_df['od_duration_value']=0.0 
flow_adelaide_df['od_distance_value']=0.0 

for idx in range(flow_adelaide_df.shape[0]):
    if idx%100 == 0:
        print(idx)

    # idx is the index in flow_adelaide_df
    origin_sa2_idx = flow_adelaide_df.loc[idx,'origin']
    destination_sa2_idx = flow_adelaide_df.loc[idx,'destination']
    
    # return the corresponding idx from OD_Google_API
    filter_idx = np.multiply(OD_google_with_map.loc[:, 'o_sa2_idx'] == origin_sa2_idx,
                             OD_google_with_map.loc[:, 'd_sa2_idx'] == destination_sa2_idx)
    idx_google_api = OD_google_with_map.index[filter_idx].tolist()[0] # this is the index in OD_google_with_map

    # 
    flow_adelaide_df.loc[idx, 'od_duration_value'] = OD_google_with_map.loc[idx_google_api, 'od_duration_value']
    flow_adelaide_df.loc[idx, 'od_distance_value'] = OD_google_with_map.loc[idx_google_api, 'od_distance_value']

# flow_adelaide_df.columns


# replace 0.0 values by 1.0
cols = ['sum_stay_duration','unique_agents','total_stays',
        'class_ART', 'class_BUS', 'class_COLL', 'class_FREE', 'class_HWY', 'class_LOCL', 
        'class_SUBA', 'class_TRK2', 'class_TRK4', 'class_UND',
        'od_duration_value', 'od_distance_value']

for col in cols:
    flow_adelaide_df.loc[flow_adelaide_df.loc[:,col] == 0.0, col] = 1.0


# dropped 433 observations. The df has nan.
flow_adelaide_df.dropna(how = 'any', inplace = True)


# add total road count as a variable
flow_adelaide_df['road_counts'] = np.sum(flow_adelaide_df[['class_ART', 'class_BUS', 'class_COLL', 'class_FREE', 'class_HWY', 'class_LOCL', 
                                         'class_SUBA', 'class_TRK2', 'class_TRK4', 'class_UND']], axis = 1)


flow_adelaide_df

flow_adelaide_df.columns

# save edge specific data
flow_adelaide_df.to_pickle('../../data_process/edge_specific_df.pickle')


# Prepare SA2 shapefile data

# origin and destination flow counts
origin_flow_counts = flow_adelaide_df.groupby(by="origin",as_index=False,sort=False).sum()[['origin','unique_agents','sum_stay_duration','total_stays']]
destination_flow_counts = flow_adelaide_df.groupby(by="destination",as_index=False,sort=False).sum()[['destination','unique_agents','sum_stay_duration','total_stays']]


# compute origin and destination entropy (w.r.t. location). flow location diversity.
# origin
origin_flow_count_n = flow_adelaide_df.groupby('origin')[['unique_agents','sum_stay_duration','total_stays']].transform('sum')
values = flow_adelaide_df[['unique_agents','sum_stay_duration','total_stays']]/origin_flow_count_n
flow_adelaide_df[['unique_agents_origin_entropy','sum_stay_duration_origin_entropy','total_stays_origin_entropy']] = \
    -(values*np.log(values))
origin_flow_entropy=flow_adelaide_df.groupby('origin',as_index=False,sort=False)[['unique_agents_origin_entropy','sum_stay_duration_origin_entropy','total_stays_origin_entropy']].sum()

# destination
destination_flow_count_n = flow_adelaide_df.groupby('destination')[['unique_agents','sum_stay_duration','total_stays']].transform('sum')
values = flow_adelaide_df[['unique_agents','sum_stay_duration','total_stays']]/destination_flow_count_n
flow_adelaide_df[['unique_agents_destination_entropy','sum_stay_duration_destination_entropy','total_stays_destination_entropy']] = \
    -(values*np.log(values))
destination_flow_entropy=flow_adelaide_df.groupby('destination',as_index=False,sort=False)[['unique_agents_destination_entropy','sum_stay_duration_destination_entropy','total_stays_destination_entropy']].sum()


origin_flow_counts

# merge data to sa2_adelaide_road_proj
# augment income and jobs
sa2_data_prep=pd.merge(sa2_adelaide_road_proj, jobs_all_sub, left_on='SA2_MAIN16', right_on='sa2_code16', how = 'inner')
sa2_data_prep=pd.merge(sa2_data_prep, origin_flow_counts, left_on='SA2_MAIN16', right_on='origin', how='inner', suffixes=[None,'_origin_counts'])
sa2_data_prep=pd.merge(sa2_data_prep, destination_flow_counts, left_on='SA2_MAIN16', right_on='destination', how='inner', suffixes=[None,'_destination_counts'])
sa2_data_prep=pd.merge(sa2_data_prep, origin_flow_entropy, left_on='SA2_MAIN16', right_on='origin', how='inner')
sa2_data_prep=pd.merge(sa2_data_prep, destination_flow_entropy, left_on='SA2_MAIN16', right_on='destination', how='inner')

# rename the '_origin_counts'
sa2_data_prep = sa2_data_prep.rename(columns={'unique_agents':'unique_agents_origin_counts',
                                      'sum_stay_duration':'sum_stay_duration_origin_counts',
                                      'total_stays':'total_stays_origin_counts'})

sa2_data_prep.columns


# save 
sa2_data_prep.to_pickle('../../data_process/node_specific_df.pickle')
