==> notebooks/jason/3.3-social-economic_var_cleaning_indigenous.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import geopandas as gpd
import geoplot
import pickle
import geoplot.crs as gcrs
import statsmodels.api as sm

mount_path = "/mnt/c/Users/jason/Dropbox (MIT)/"
#mount_path = "/Users/shenhaowang/Dropbox (MIT)/project_media_lab_South_Australia/"

social_economic_df = pd.read_csv(mount_path + "/SA data/NATSEM_-_Social_and_Economic_Indicators_-_Indigenous_Indicators_SA2_2016.csv/data6735501229514139095.csv")


social_economic_df = social_economic_df.dropna()

social_economic_df.columns

social_economic_df.rename(columns= {" sa2_code16":"SA2_MAIN16"},inplace=True)

helper = [str(elt) for elt in social_economic_df["SA2_MAIN16"].values]

helper

social_economic_df["SA2_MAIN16"] = helper

social_economic_df.head()

social_economic_df.shape

node_degree_entropy_df = pd.read_pickle('../../data_process/node_degree_entropy_df.pickle')

node_degree_entropy_df = node_degree_entropy_df.merge(social_economic_df, how="left", on="SA2_MAIN16")

node_degree_entropy_df

node_degree_entropy_df.shape

node_degree_entropy_df = node_degree_entropy_df.dropna()

node_degree_entropy_df.shape

node_degree_entropy_df.to_pickle("../../data_process/node_degree_entropy_social_indigenous_df.pickle")
