==> notebooks/jason/9_gravity_model.ipynb <==
# Gravity Model

This script build a basic gravity model for analysis. 

Exemplary paper: Economic Outcome Predicted by Diversity

Q: Need economic indicators (e.g. GDP) and jobs. 

Q: Unresolved - Traffic assignment problem.


pwd

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import geopandas as gpd
import geoplot

jobs_all = pd.read_csv("../data/SA2_Jobs_All_Jobs_and_Income.csv")
jobs_industries = pd.read_csv("../data/SA2_Jobs_In_Australia_Employee_Jobs_and_Income.csv")

jobs_all

jobs_industries

new_idx = []

for col in jobs_industries.columns:
    if col[0]==' ':
        new_idx.append(col[1:])
    else:
        new_idx.append(col)

jobs_industries.columns = new_idx
jobs_industries.columns

new_idx = []

for col in jobs_all.columns:
    if col[0]==' ':
        new_idx.append(col[1:])
    else:
        new_idx.append(col)

jobs_all.columns = new_idx
jobs_all.columns


jobs_all['sa2_code16'] = jobs_all['sa2_code16'].astype('str')


# notes:
# 'sa2_code16'


# read shapefile. 
sa2_shape = gpd.read_file("../data/sa2/SA2_2016_AUST.shp")
# Keep only the Adelaide area
adelaide_sa4_set=['401','402','403','404']
sa2_adelaide = sa2_shape.loc[sa2_shape.SA4_CODE16.isin(adelaide_sa4_set)]
print(sa2_adelaide.shape)
print(sa2_adelaide.head(10))


# merge
sa2 = pd.merge(sa2_adelaide, jobs_all, left_on='SA2_MAIN16', right_on='sa2_code16', how='inner')


sa2 # only the Adelaide area


# Model

Gravity model:
$$ T_{ij} = \frac{T_i^{\alpha} T_j^{\beta}}{C_{ij}^{\gamma}}$$

Log-linear form:
$$ \log T_{ij} = \alpha \log T_i + \beta \log T_j - \gamma \log C_{ij} $$

Let us use basic economic info here.
1. $T_{ij}$ is the mobility flow from i to j.
2. $T_i$ and $T_j$ represent attractions, e.g. number of total jobs.
3. $C_{ij}$ represents frictions, e.g. travel time or other road attributes. Ideally, $C_{ij}$ can represent the actionable variables, such as road capacity/types. 

OLS can be used to estimate the three parameters $\alpha$, $\beta$, and $\gamma$.

More generally, we can write the gravity model as:
$$ \log T_{ij} = f_1(\log T_i, \log T_j; w_1) + f_2(\log C_{ij}; w_2) $$
